refactor(clients.snssqs): rename module snssqs_client -> snssqs
