+++ src/tomodachi_testcontainers/clients/__init__.py
@@ -1,4 +1,4 @@
-from tomodachi_testcontainers.clients.snssqs_client import SNSSQSTestClient
+from tomodachi_testcontainers.clients.snssqs import SNSSQSTestClient
 
 __all__ = [
     "SNSSQSTestClient",
